+++ src/Symfony/Component/Asset/Tests/PackagesTest.php
@@ -51,7 +51,7 @@ class PackagesTest extends TestCase
     {
         $packages = new Packages(
             new Package(new StaticVersionStrategy('default')),
-            ['a' => new Package(new StaticVersionStrategy('a'))]
+            new \ArrayIterator(['a' => new Package(new StaticVersionStrategy('a'))])
         );
 
         $this->assertSame('/foo?default', $packages->getUrl('/foo'));